pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
